Moves `StoryboardBackgroundSprite` and `DrawableStoryboardBackgroundSprite` to its own files.
